+++ src/datasources/results/components/variable-editors/ResultsVariableQueryEditor.tsx
@@ -19,6 +19,7 @@ import {
 import { ResultsDataSource } from 'datasources/results/ResultsDataSource';
 import { StepsQueryBuilderWrapper } from '../query-builders/steps-querybuilder-wrapper/StepsQueryBuilderWrapper';
 import { TAKE_LIMIT } from 'datasources/results/constants/QuerySteps.constants';
+import { FloatingError } from 'core/errors';
 
 type Props = QueryEditorProps<ResultsDataSource, ResultsQuery, ResultsDataSourceOptions>;
 
@@ -205,7 +206,7 @@ export function ResultsVariableQueryEditor({ query, onChange, datasource }: Prop
             labelWidth={26}
             tooltip={tooltips.productName}
             invalid={!isProductSelectionInStepsValid}
-            error="This field requires at least one product to be selected.">
+            error="You must select at least one product in this field.">
             <MultiSelect
               maxVisibleValues={5}
               width={65}
@@ -247,6 +248,7 @@ export function ResultsVariableQueryEditor({ query, onChange, datasource }: Prop
           </InlineField>
         </>
       )}
+      <FloatingError message={queryResultsDataSource.current.errorTitle} innerMessage={queryResultsDataSource.current.errorDescription} severity='warning'/>
     </>
   );
 }